Avoid explicit eval in favor of generated blocks
I had some errors thrown while precompiling because of something to do
with the `eval` calls. A more common pattern for the code that was using
`eval` is to use a `@generated` block in a function. Generated functions
have access to the **types** of args passed to it, and can generate code
blocks based on those types; which happens to be exactly what the
`impl_` functions were doing.

With this, I'm able to connect to a local ClickHouse instance running on
my local computer.